# docs/swagger.yaml
basePath: /api
definitions:
...
  tagshttp.createTagRequest:
    properties:
      board_id:
        type: string
      color:
        type: string
      name:
        type: string
    required:
    - board_id
    - color
    - name
    type: object
  tagshttp.tagResponse:
    properties:
      board_id:
        type: string
      color:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  tagshttp.updateTagRequest:
    properties:
      color:
        type: string
      name:
        type: string
    type: object
...
host: localhost:8003
info:
  contact: {}
  description: API for managing kanban boards
  title: Student Kanban API
  version: "1.0"
paths:
...
  /cards/{id}/tags:
    get:
      description: Returns all tags attached to a given card
      parameters:
      - description: Card UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/tagshttp.tagResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all tags for a card
      tags:
      - tags
  /cards/{id}/tags/{tag_id}:
    delete:
      description: Removes a tag from a card
      parameters:
      - description: Card UUID
        in: path
        name: id
        required: true
        type: string
      - description: Tag UUID
        in: path
        name: tag_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Detach a tag from a card
      tags:
      - tags
    post:
      description: Attaches an existing tag to a card
      parameters:
      - description: Card UUID
        in: path
        name: id
        required: true
        type: string
      - description: Tag UUID
        in: path
        name: tag_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Attach a tag to a card
      tags:
      - tags
...
  /tags:
    post:
      consumes:
      - application/json
      description: Creates a new tag for a board
      parameters:
      - description: Tag data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/tagshttp.createTagRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/tagshttp.tagResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new tag
      tags:
      - tags
  /tags/{id}:
    delete:
      description: Deletes a tag by its UUID
      parameters:
      - description: Tag UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a tag
      tags:
      - tags
    get:
      description: Returns a single tag by its UUID
      parameters:
      - description: Tag UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tagshttp.tagResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a tag by ID
      tags:
      - tags
    put:
      consumes:
      - application/json
      description: Updates a tag's name and/or color. At least one field must be provided
      parameters:
      - description: Tag UUID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/tagshttp.updateTagRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tagshttp.tagResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a tag
      tags:
      - tags
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"